Initial package commit.
This is a generated package plus some changes and additions. Mainly
focused on working with Authentication. All the logic with OIDC and SSS
is running well enough. The last needed piece is to set up the local
authentication cookie and hook the authentication logic with it. Then it
would be possible to fucuse on creating discussions.