Changed router logic, always include targetEntityUri.
The app will expect to get the TargetEntityUri to be provided into every
route, this way we keep track of things. Every route will also silently
set it to a sorage on to of $rootScope. This is needed for the
reauthentication step if OIDC expires. All yhe main logic should be
fully functional now.