Added logic of answers to discussion.
The whole thing needs a rewrite as there currently is a huge portion of
repetitive code (use the copy-paste magic as a shortcut). WIll also
chnage logic for loading data on files attached to entitites, used in
downloading. Will use a service and be loading as someone clicks, this
information does not change and could be centrally stored and reused by
multiple controllers, this should also take care about not making
repatitions.