Added a way to edit entries, using modal.
There is a code repetition in this case as a lot of it is almost the
same or very similar to the DiscussionEditModal controller, directive
and form. Although changes are substantial and there is not enough time
to rewrite it in a better way. This would still need some testing to
make sure everything is working as expected.