Added a check to filling lookup table.
The code will check for loading failure and display a danger or warning
message, the latter one is displayed in case no versions could be loaded
and the target might not be an episode. Later on there could be one more
loading call added to chec what type of entity we are dealing with, but
currently it seems sufficient.