DT-6 #time 1h Changed REST API service locations.
Introduced a new constant and added it to all services. This would still
need to be tested with new version of SSS. It might be beneficial to
create a service taht would deal with REST endpoint (single service)
Base URL construction. Then it could be reused in other services instead
of including the config and sssRestPrefix and constructing manually.
Although it is not highly needed (especially with uncertainty of DT
being outdated and functionality becoming part of each tool).