DT-7 #resolve Configurator should work everywhere.
Rewrote the confgurator script to work on: Debian/Ubuntu, RedHat/CentOS
and OS X (tested with all of those). Removed any Bash specific parts and
replaced those with something that works in all of these distributions
by default. NB! There is an issue with running that with either pure
Bash or in RedHat/CentOS - colors will not be shown, color codes will be
cluttering the texts instead. This could be resolved with using -e flag
on Echo command, although one has to be sure that it is available (this
might be tackled later as it is not highly important).