DT-8 LL-1274 Fixed an issue with a few logs not being sent.
This wsas due to the autheticate logic still running and used not being
authenticated, when initially entering the app. All the rest of
navigation still preserves the state of the app that was before. Added
an event to be sent after authentication really does happen. The code
then catches it and runs the routines as needed. The logic for case of
user being logged in is still the same.