DT-13 DT-12 DT-17 DT-16 #resolve All chnaged added.
All the changes should be in, the code might not be fully finalized, but
it is almost there. Document is attached to the Episode instead, but it
makes not difference as the connection creates is two-way. All needs to
be tested with the LD test insatnce that is soon to become available.