D-52 #close, LL-1352 #close, added collaborators and more.
In addition to collaborators, a filter called userLabel was added. That
one will be applied to all the cases showing user labels and would
remove the @whatever email address endings (making it consistent with
what BnP shows). The collaborator data will be shown as a popover when
moving over BnP button. The could would also be displayed within that
(that is not consistent with BnP, but there does not seem to be any more
free space witin the navbar on smaller screens - it just becomes
multilined).